requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
mistralai>=1.0.0
python-dotenv>=1.0.0
elevenlabs>=1.0.0
